requirements.txt
jenkins-webapi
BlinkStick
requests
